Update README.md
I don't know if this will eliminate confusion. I was accurate before I just assumed I knew what it was going to tell me.
The next line: "You can include `/js/boundarizr.js` (or .min.js if you prefer) and call `Boundarizr.testCurrentDocument()`."
is not related to the bookmarklet right? It's about adding it to a project?