Add class name(s) to the boundary's label.
Since id usage for elements is less common than classes, the label for
the layout boundary is often “div”, which is not helpful. This change
will make the label similar to how the dev tools show elements
`div#id.class-one.class-two`

I also ran build.js to update the bookmarklet and minified file.